Moved some MySQL adapter tests to their own method cests.
