+++ tests/integration/Db/Adapter/Pdo/MysqlCest.php
@@ -15,7 +15,6 @@ use Codeception\Example;
 use IntegrationTester;
 use PDOException;
 use Phalcon\Db\Adapter\Pdo\Mysql;
-use Phalcon\Db\Reference;
 use Phalcon\Test\Integration\Db\Dialect\Helper\MysqlHelper;
 
 class MysqlCest extends MysqlHelper
@@ -43,118 +42,6 @@ class MysqlCest extends MysqlHelper
                 'Unable to connect to the database: ' . $e->getMessage()
             );
         }
-    }
-
-    /**
-     * Tests Mysql::listTables
-     *
-     * @author Phalcon Team <team@phalconphp.com>
-     * @since  2016-08-03
-     */
-    public function testListTables(IntegrationTester $I)
-    {
-        $expected = [
-            'albums',
-            'artists',
-            'childs',
-            'customers',
-            'dialect_table',
-            'dialect_table_intermediate',
-            'dialect_table_remote',
-            'foreign_key_child',
-            'foreign_key_parent',
-            'identityless_requests',
-            'issue12071_body',
-            'issue12071_head',
-            'issue_11036',
-            'issue_1534',
-            'issue_2019',
-            'm2m_parts',
-            'm2m_robots',
-            'm2m_robots_parts',
-            'package_details',
-            'packages',
-            'parts',
-            'personas',
-            'personnes',
-            'ph_select',
-            'prueba',
-            'robots',
-            'robots_parts',
-            'songs',
-            'stats',
-            'stock',
-            'subscriptores',
-            'table_with_string_field',
-            'tipo_documento',
-            'users',
-        ];
-
-
-        $I->assertEquals(
-            $expected,
-            $this->connection->listTables()
-        );
-
-
-        $dbName = env('DATA_MYSQL_NAME', 'phalcon_test');
-
-        $I->assertEquals(
-            $expected,
-            $this->connection->listTables($dbName)
-        );
-    }
-
-    /**
-     * Tests Mysql::describeReferences
-     *
-     * @author Wojciechj Ślawski <jurigag@gmail.com>
-     * @since  2016-09-28
-     */
-    public function testDescribeReferencesColumnsCount(IntegrationTester $I)
-    {
-        $actual = $this->connection->describeReferences(
-            'robots_parts',
-            env('DATA_MYSQL_NAME')
-        );
-
-        $I->assertCount(2, $actual);
-
-
-        $I->assertCount(
-            2,
-            $this->connection->describeReferences('robots_parts', null)
-        );
-
-
-        $references = $actual;
-
-        /** @var Reference $reference */
-        foreach ($references as $reference) {
-            $I->assertCount(
-                1,
-                $reference->getColumns()
-            );
-        }
-    }
-
-    /**
-     * Tests Mysql::escapeIdentifier
-     *
-     * @author       Sid Roberts <https://github.com/SidRoberts>
-     * @since        2016-11-19
-     *
-     * @dataProvider testEscapeIdentifierProvider
-     */
-    public function testEscapeIdentifier(IntegrationTester $I, Example $example)
-    {
-        $identifier = $example['identifier'];
-        $expected   = $example['expected'];
-
-        $I->assertEquals(
-            $expected,
-            $this->connection->escapeIdentifier($identifier)
-        );
     }
 
     /**
@@ -219,31 +106,6 @@ class MysqlCest extends MysqlHelper
                 $example['sql']
             )
         );
-    }
-
-    private function testEscapeIdentifierProvider(): array
-    {
-        return [
-            [
-                'identifier' => 'robots',
-                'expected'   => '`robots`',
-            ],
-
-            [
-                'identifier' => ['schema', 'robots'],
-                'expected'   => '`schema`.`robots`',
-            ],
-
-            [
-                'identifier' => '`robots`',
-                'expected'   => '```robots```',
-            ],
-
-            [
-                'identifier' => ['`schema`', 'rob`ots'],
-                'expected'   => '```schema```.`rob``ots`',
-            ],
-        ];
     }
 
     private function shouldAddForeignKeyProvider(): array
